backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
python-multipart==0.0.9
slowapi==0.1.9
openai==1.50.0
PyPDF2==3.0.1
python-docx==1.1.2
aiofiles==24.1.0
python-dotenv==1.0.1